Worked on SonarCloud findings

+++ Renderer/src/net/retrocarnage/editor/renderer/common/SpritePainter.java
@@ -104,9 +104,9 @@ public class SpritePainter {
         final int h = (int) Math.floor(image.getHeight() * cos + image.getWidth() * sin);
         final BufferedImage rotatedImage = new BufferedImage(w, h, image.getType());
         final AffineTransform at = new AffineTransform();
-        at.translate(w / 2, h / 2);
+        at.translate(w / (double) 2, h / (double) 2);
         at.rotate(rads, 0, 0);
-        at.translate(-image.getWidth() / 2, -image.getHeight() / 2);
+        at.translate(-image.getWidth() / (double) 2, -image.getHeight() / (double) 2);
         final AffineTransformOp rotateOp = new AffineTransformOp(at, AffineTransformOp.TYPE_BILINEAR);
         rotateOp.filter(image, rotatedImage);
         return rotatedImage;
